Implemented looping around wordsearch words
This has 3 additional transparent layers to use to display the images in
case of overlapping.  * This may not be enough layers in certain
situations, but more could be added in to ensure there will be no
glitches.  (But it depends on how words are allowed to overlap each
other- presumably there could never be more than 8 overlaps?!)